update authsig with new permissions interface
